+++ STACpopulator/implementations/RDPS_CRIM/add_RDPS.py
@@ -13,7 +13,6 @@ class RDPSpopulator(THREDDSPopulator):
     name = "RDPS_CRIM"
     description = "RDPS STAC populator from a THREDDS catalog or NCML XML."
     data_model = RDPSDataModel
-    item_geometry_model = None  # Unnecessary, but kept for consistency
 
     def create_stac_item(self, item_name: str, item_data: dict[str, Any]) -> dict[str, Any]:
         """Return a STAC item."""