+++ CHANGES.md
@@ -10,6 +10,8 @@
 * Rename `populator_base.py` to `populators.py` since it now contains more than just the base populator
 * No longer require new implementations to list populator modules in `implementations.__init__.py`
 * Implementation discovery is now done by keeping track of concrete implementations of the STACpopulatorBase class
+* Add ability to parse vertical data from CF metadata extracted from NCML files
+* Convert longitude values in range 0-360 degrees to -180-180 degrees to comply with WGS84
 
 ## [0.12.0](https://github.com/crim-ca/stac-populator/tree/0.12.0) (2025-11-20)